refactoring code to follow correct pattern

+++ models/Job.go
@@ -1,7 +1,7 @@
 package models
 
 type Job struct {
-	JobID        int64     `json:"JobID"`
+	JobID        int32     `json:"JobID"`
 	JobTitle     string    `json:"JobTitle"`
 	Salary       float32   `json:"Salary"`
 }
